new/requirements.txt
openai
elasticsearch
elasticsearch[helpers]
chromadb
voyageai
python-dotenv
streamlit
pymongo
opik